Moved DiagonalMatrix and IdentityMatrix to group Matrices and Linear Algebra
