+++ mathics/builtin/tensors.py
@@ -15,7 +15,6 @@ from mathics.builtin.base import Builtin, BinaryOperator
 from mathics.core.expression import Expression
 from mathics.core.atoms import (
     Integer,
-    Integer0,
     String,
 )
 from mathics.core.symbols import (
@@ -169,38 +168,6 @@ class ArrayQ(Builtin):
         return SymbolTrue
 
 
-class DiagonalMatrix(Builtin):
-    """
-    <dl>
-      <dt>'DiagonalMatrix[$list$]'
-      <dd>gives a matrix with the values in $list$ on its diagonal and zeroes elsewhere.
-    </dl>
-
-    >> DiagonalMatrix[{1, 2, 3}]
-     = {{1, 0, 0}, {0, 2, 0}, {0, 0, 3}}
-    >> MatrixForm[%]
-     = 1   0   0
-     .
-     . 0   2   0
-     .
-     . 0   0   3
-
-    #> DiagonalMatrix[a + b]
-     = DiagonalMatrix[a + b]
-    """
-
-    def apply(self, list, evaluation):
-        "DiagonalMatrix[list_List]"
-
-        result = []
-        n = len(list.leaves)
-        for index, item in enumerate(list.leaves):
-            row = [Integer0] * n
-            row[index] = item
-            result.append(Expression("List", *row))
-        return Expression("List", *result)
-
-
 class Dimensions(Builtin):
     """
     <dl>
@@ -268,22 +235,6 @@ class Dot(BinaryOperator):
     }
 
     summary_text = "dot product"
-
-
-class IdentityMatrix(Builtin):
-    """
-    <dl>
-    <dt>'IdentityMatrix[$n$]'
-        <dd>gives the identity matrix with $n$ rows and columns.
-    </dl>
-
-    >> IdentityMatrix[3]
-     = {{1, 0, 0}, {0, 1, 0}, {0, 0, 1}}
-    """
-
-    rules = {
-        "IdentityMatrix[n_Integer]": "DiagonalMatrix[Table[1, {n}]]",
-    }
 
 
 class Inner(Builtin):
